Refactor methods for building the order request into separate class
