+++ GeeksCoreLibrary/Modules/Payments/Services/MollieService.cs
@@ -18,6 +18,7 @@ using GeeksCoreLibrary.Core.Models;
 using GeeksCoreLibrary.Modules.Databases.Interfaces;
 using GeeksCoreLibrary.Modules.Objects.Interfaces;
 using GeeksCoreLibrary.Modules.Payments.Enums;
+using GeeksCoreLibrary.Modules.Payments.Helpers.Mollie;
 using GeeksCoreLibrary.Modules.Payments.Interfaces;
 using GeeksCoreLibrary.Modules.Payments.Models;
 using GeeksCoreLibrary.Modules.Payments.Models.Mollie;
@@ -84,7 +85,12 @@ namespace GeeksCoreLibrary.Modules.Payments.Services
             mollieSettings.Locale = await objectsService.FindSystemObjectByDomainNameAsync("MOLLIE_locale");
 
             var mollieClient = new OrderClient(mollieSettings.ApiKey);
-            var orderRequest = await CreateOrderRequestAsync(totalPrice, mollieSettings, invoiceNumber, paymentMethodSettings, shoppingBaskets, userDetails);
+            var orderRequestBuilder = new OrderRequestBuilder(
+                shoppingBasketsService, 
+                mollieSettings, 
+                gclSettings, 
+                paymentMethodSettings);
+            var orderRequest = await orderRequestBuilder.CreateOrderRequestAsync(totalPrice, invoiceNumber, shoppingBaskets, userDetails);
 
             OrderResponse orderResponse;
             try
@@ -132,36 +138,6 @@ namespace GeeksCoreLibrary.Modules.Payments.Services
 
                 await shoppingBasketsService.SaveAsync(main, lines, await shoppingBasketsService.GetSettingsAsync());
             }
-        }
-
-        private static string GetIssuerName(string issuerValue)
-        {
-            var issuerConstants = typeof(IdealIssuers).GetFields(BindingFlags.Public | BindingFlags.Static);
-            var issuerConstant = issuerConstants.FirstOrDefault(mi => mi.Name.Equals(issuerValue, StringComparison.OrdinalIgnoreCase) || mi.Name.Equals($"ideal_{issuerValue}", StringComparison.OrdinalIgnoreCase));
-
-            if (issuerConstant != null)
-            {
-                return (string)issuerConstant.GetValue(null);
-            }
-
-            return null;
-        }
-
-        private string BuildUrl(string webhookUrl, string invoiceNumber)
-        {
-            if (httpContextAccessor?.HttpContext == null)
-            {
-                return String.Empty;
-            }
-            
-            // TODO: Refactor this method so that we can use it for all PSPs.
-            var webhookUrlBuilder =  new UriBuilder(webhookUrl);
-            var queryString = HttpUtility.ParseQueryString(webhookUrlBuilder.Query);
-            queryString["invoice_number"] = invoiceNumber;
-
-            webhookUrlBuilder.Query = queryString.ToString() ?? String.Empty;
-
-            return webhookUrlBuilder.ToString();
         }
 
         /// <inheritdoc />
@@ -297,150 +273,6 @@ namespace GeeksCoreLibrary.Modules.Payments.Services
             }
 
             return totalPrice;
-        }
-
-        private async Task<OrderRequest> CreateOrderRequestAsync(decimal totalPrice,
-            MollieSettingsModel mollieSettings,
-            string invoiceNumber,
-            PaymentMethodSettingsModel paymentMethodSettings,
-            ICollection<(WiserItemModel Main, List<WiserItemModel> Lines)> shoppingBaskets,
-            WiserItemModel userDetails)
-        {
-            var orderRequest = new OrderRequest
-            {
-                Amount = new Amount
-                {
-                    Value = totalPrice.ToString("F2", CultureInfo.InvariantCulture),
-                    Currency = mollieSettings.Currency
-                },
-                OrderNumber = invoiceNumber,
-                RedirectUrl = BuildUrl(mollieSettings.ReturnUrl, invoiceNumber),
-                WebhookUrl = BuildUrl(mollieSettings.WebhookUrl, invoiceNumber),
-                Locale =  mollieSettings.Locale,
-                Method = paymentMethodSettings.ExternalName,
-                Lines = await ConvertShoppingBasketsToOrderLinesAsync(shoppingBaskets, mollieSettings),
-                BillingAddress = CreateBillingAddress(userDetails),
-                ShippingAddress = CreateAddress(userDetails, "shipping_"),
-                Testmode = gclSettings.Environment.InList(Environments.Test, Environments.Development)
-            };
-
-            if (String.Equals(paymentMethodSettings?.ExternalName, "ideal", StringComparison.OrdinalIgnoreCase))
-            {
-                var issuerValue = shoppingBaskets.First().Main.GetDetailValue(Components.OrderProcess.Models.Constants.PaymentMethodIssuerProperty);
-                var issuerName = GetIssuerName(issuerValue);
-                orderRequest.Payment = new IDealSpecificParameters
-                {
-                    Issuer = issuerName,
-                };
-            }
-            
-            // Metadata is always sent back.
-            orderRequest.Metadata = invoiceNumber;
-            return orderRequest;
-        }
-        
-        private OrderAddressDetails CreateAddress(WiserItemModel userDetails, string detailKeyPrefix = "")
-        {
-            //If a prefix is given but any of the required values doesn't contain a value return null.
-            if (!String.IsNullOrWhiteSpace(detailKeyPrefix) && 
-                (String.IsNullOrWhiteSpace(userDetails.GetDetailValue($"{detailKeyPrefix}street")))
-                 || String.IsNullOrWhiteSpace(userDetails.GetDetailValue($"{detailKeyPrefix}zipcode"))
-                 || String.IsNullOrWhiteSpace(userDetails.GetDetailValue($"{detailKeyPrefix}city"))
-                 || String.IsNullOrWhiteSpace(userDetails.GetDetailValue($"{detailKeyPrefix}country")))
-            {
-                return null;
-            }
-
-            return new OrderAddressDetails()
-            {
-                StreetAndNumber =
-                    $"{userDetails.GetDetailValue($"{detailKeyPrefix}street")} {userDetails.GetDetailValue($"{detailKeyPrefix}housenumber")}{userDetails.GetDetailValue($"{detailKeyPrefix}housenumber_suffix")}",
-                PostalCode = userDetails.GetDetailValue($"{detailKeyPrefix}zipcode"),
-                City = userDetails.GetDetailValue($"{detailKeyPrefix}_city"),
-                Country = userDetails.GetDetailValue($"{detailKeyPrefix}_country"),
-            };
-        }
-
-        private OrderAddressDetails CreateBillingAddress(WiserItemModel userDetails)
-        {
-            var address = CreateAddress(userDetails);
-            address.OrganizationName = userDetails.GetDetailValue("companyname");
-            address.GivenName = userDetails.GetDetailValue("firstname");
-            address.FamilyName = userDetails.GetDetailValue("lastname");
-            address.Email = userDetails.GetDetailValue("email");
-            address.Phone = userDetails.GetDetailValue("phone");
-            return address;
-        }
-
-        private async Task<IEnumerable<OrderLineRequest>> ConvertShoppingBasketsToOrderLinesAsync(
-            ICollection<(WiserItemModel Main, List<WiserItemModel> Lines)> shoppingBaskets,
-            MollieSettingsModel mollieSettings)
-        {
-            var basketSettings = await shoppingBasketsService.GetSettingsAsync();
-            var orderRequests = new List<OrderLineRequest>();
-            foreach (var basket in shoppingBaskets)
-            {
-                foreach (var basketLine in basket.Lines)
-                {
-                    OrderLineRequest orderLineRequest =
-                        await ConvertOrderLineFromBasketLineAsync(
-                            basket.Main, 
-                            basketLine,
-                            basketSettings,
-                            mollieSettings);
-                    orderRequests.Add(orderLineRequest);
-                }
-            }
-
-            return orderRequests;
-        }
-
-        private async Task<OrderLineRequest> ConvertOrderLineFromBasketLineAsync(
-            WiserItemModel basket,
-            WiserItemModel basketLine,
-            ShoppingBasketCmsSettingsModel basketSettings,
-            MollieSettingsModel mollieSettings)
-        {
-            var linePrice = await shoppingBasketsService.GetLinePriceAsync(
-                basket,
-                basketLine,
-                basketSettings,
-                ShoppingBasket.PriceTypes.PspPriceInVat
-            );
-            var linePriceVatOnly = await shoppingBasketsService.GetLinePriceAsync(
-                basket,
-                basketLine,
-                basketSettings,
-                ShoppingBasket.PriceTypes.VatOnly
-            );
-            var lineProductQuantity = Convert.ToInt32(basketLine.GetDetailValue(basketSettings.QuantityPropertyName));
-
-            return new OrderLineRequest()
-            {
-                Name = basketLine.GetDetailValue("title"),
-                UnitPrice = new Amount()
-                {
-                    Value = (linePrice / lineProductQuantity).ToString("F2", CultureInfo.InvariantCulture),
-                    Currency = mollieSettings.Currency
-                },
-                TotalAmount = new Amount()
-                {
-                    Value = linePrice.ToString("F2", CultureInfo.InvariantCulture),
-                    Currency = mollieSettings.Currency
-                },
-                DiscountAmount = new Amount()
-                {
-                    Value = basketLine.GetDetailValue(basketSettings.DiscountPropertyName),
-                    Currency = mollieSettings.Currency
-                },
-                VatAmount = new Amount()
-                {
-                    Value = linePriceVatOnly.ToString("F2", CultureInfo.InvariantCulture),
-                    Currency = mollieSettings.Currency
-                },
-                Quantity = lineProductQuantity,
-                VatRate = basketLine.GetDetailValue(basketSettings.VatRatePropertyName),
-            };
         }
     }
 }
